Fix nation deletion listener using PreDeleteNationEvent.
This PR will make it so the nation refund is available under more types
of deletion. Single-town nations being deleted were likely to not
register the king properly. By using the PreDeleteNationEvent, on
monitor priority with canceled events ignored, we will be able to get
the King (probably) every time.