Fix trap mitigation explosion handling
#### Description: 

`event.setBlockList(finalExplodeList);` was contained within the `if` statement that handles explosions for the Cannons integration, so `TownyExplodingBlocksEvent` would not be modified if trap mitigation was enabled but the Cannons integration was disabled. 

#### New Nodes/Commands/ConfigOptions: 
None.

#### Relevant Issue ticket: 
None.

[x] I have tested this pull request for defects on a server.

By making this pull request, I represent that I have the right to waive copyright and related rights to my contribution and agree that all copyright and related rights in my contributions are waived, and I acknowledge that the TownyAdvanced organization has the copyright to use and modify my contribution under the SiegeWar [License](https://github.com/TownyAdvanced/SiegeWar/blob/master/LICENSE.md) for perpetuity.