Add SiegeCamp object, created when a Siege is about to start.
Object collects all the information required for the
PreSiegeWarStartEvent and starting the actual Siege.

Upon creation, a SiegeCamp begins a 10 minute period where on each
minute if the attackers are present within the TimedPointZone, the
attackers gain 7 points. After the 10 minute period, if the Attackers
have 50 points the Siege will begin.

Closes #402.

TODO:
- [ ] Language strings.
- [ ] Configuration section for points, times.
- [ ] Basic testing.