Translate our translatable messages more thoroughly.
PR largely removes the usage of Translation, which defaults to the
server default locale, replacing it with Translatable where possible.

Exceptions:
- EventListeners do not always deliver a player that will be shown the
cancel message.
- Dynmap/Bossbar/SiegeType areas where we actually do want things to be
using the server default locale, for universal messages without
individual readers.

Additionally, many areas where we caught TownyExceptions we were not
choosing to show the message in the player's locale.

Fixes a thing in the BannerControlUtil that would cause an evaluation to
fail before completion.