Make it possible to show the time to the next BattleSession from more than 1 day in advance.
Sessions might not be starting an extended period of time, this PR makes
it possible for the code to look more than 1 day in advance when it
cannot otherwise find a next session for the day.

It will look up to 6 days away, since we don't want to look at the
current day of the week.